feat: finalizando cadastro de clientes e localizacoes

+++ server/src/api/resources/addresses/addresses.php
@@ -1,5 +1,5 @@
 <?php
-require_once './../../../config.php';
+include_once('./../../../app/addresses/AddressesController.php');
 
 spl_autoload_register(function ($className) {
   if (file_exists("{$className}.php")) {
@@ -7,18 +7,11 @@ spl_autoload_register(function ($className) {
   }
 });
 
-use src\App\Utils\Utils;
 use src\App\Controllers\AddressesController;
 
-Utils::buildRoute("/addresses/get-all", "GET", function () {
-  header("content-type: application/json");
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
   $addresses = new AddressesController();
-  echo json_encode($addresses->getAllAddresses());
-});
-
-Utils::buildRoute("/addresses/new-address", "POST", function () {
-  $addresses = new AddressesController();
-  $addresses->setCustomerId($_POST["customerId"]);
+  $addresses->setCustomerCPF($_POST["customerCPF"]);
   $addresses->setAddressName($_POST["addressName"]);
   $addresses->setStreet($_POST["addressStreet"]);
   $addresses->setNumber($_POST["addressNumber"]);
@@ -28,17 +21,11 @@ Utils::buildRoute("/addresses/new-address", "POST", function () {
   $addresses->settAddressState($_POST["addressState"]);
   $addresses->setZipCode($_POST["addressZipCode"]);
   echo $addresses->newAddress();
-});
+}
 
-Utils::buildRoute("/addresses/get-one/:id", "GET", function () {
-  header("content-type: application/json");
+if ($_SERVER['REQUEST_METHOD'] === 'PUT') {
   $addresses = new AddressesController();
-  echo json_encode($addresses->getOneAddress($_GET["addressId"]));
-});
-
-Utils::buildRoute("/addresses/update/:addressId", "POST", function () {
-  $addresses = new AddressesController();
-  $addresses->setCustomerId($_POST["customerId"]);
+  $addresses->setCustomerCPF($_POST["customerCPF"]);
   $addresses->setAddressName($_POST["addressName"]);
   $addresses->setStreet($_POST["addressStreet"]);
   $addresses->setNumber($_POST["addressNumber"]);
@@ -48,20 +35,30 @@ Utils::buildRoute("/addresses/update/:addressId", "POST", function () {
   $addresses->settAddressState($_POST["addressState"]);
   $addresses->setZipCode($_POST["zipCode"]);
   echo $addresses->updateAddress($_GET["addressId"]);
-});
+}
 
-Utils::buildRoute("/addresses/delete/:addressId", "GET", function () {
+if ($_SERVER['REQUEST_METHOD'] === 'GET' && !isset($_GET['addressId'])) {
+  $addresses = new AddressesController();
+  echo json_encode($addresses->getAllAddresses());
+}
+
+if ($_SERVER['REQUEST_METHOD'] === 'GET' && isset($_GET['addressId'])) {
+  $addresses = new AddressesController();
+  echo json_encode($addresses->getOneAddress($_GET["addressId"]));
+}
+
+if ($_SERVER['REQUEST_METHOD'] === 'DELETE' && isset($_GET['addressId'])) {
   $addresses = new AddressesController();
   echo $addresses->deleteAddress($_GET["addressId"]);
-});
+};
 
-Utils::buildRoute("/addresses/get-customer-addresses/:id", "GET", function () {
+if ($_SERVER['REQUEST_METHOD'] === 'GET' && isset($_GET['customerId'])) {
   $addresses = new AddressesController();
   echo json_encode($addresses->getAddressesByCustomerId($_GET["customerId"]));
-});
+};
 
-Utils::buildRoute("/addresses/delete-customer-address/:id", "GET", function () {
+if ($_SERVER['REQUEST_METHOD'] === 'DELETE' && isset($_GET['customerId'])) {
   $addresses = new AddressesController();
   echo $addresses->deleteAddressByCustomerId($_GET["customerId"]);
-});
+};
 ?>
