+++ server/src/api/resources/users/users.php
@@ -1,5 +1,5 @@
 <?php
-require_once './../../../config.php';
+include_once('./../../../app/users/UsersController.php');
 
 spl_autoload_register(function ($className) {
   if (file_exists("{$className}.php")) {
@@ -10,42 +10,40 @@ spl_autoload_register(function ($className) {
 use src\App\Utils\Utils;
 use src\App\Controllers\UsersController;
 
-Utils::buildRoute("/users/login", "GET", function () {
-  $users = new UsersController();
-  $users->setUserEmail($_POST["userEmail"]);
-  $users->setUserPassword($_POST["userPassword"]);
-});
+// Utils::buildRoute("/users/login", "GET", function () {
+//   $users = new UsersController();
+//   $users->setUserEmail($_POST["userEmail"]);
+//   $users->setUserPassword($_POST["userPassword"]);
+// });
 
-Utils::buildRoute("/users/get-all", "GET", function () {
-  header("content-type: application/json");
+if ($_SERVER['REQUEST_METHOD'] === 'GET' && !isset($_GET['userId'])) {
   $users = new UsersController();
   echo json_encode($users->getAllUsers());
-});
+}
 
-Utils::buildRoute("/users/create-user", "POST", function () {
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
   $users = new UsersController();
   $users->setUserName($_POST["userName"]);
   $users->setUserEmail($_POST["userEmail"]);
   $users->setUserPassword($_POST["userPassword"]);
   echo $users->newUser();
-});
+}
 
-Utils::buildRoute("/users/get-one/:userId", "GET", function () {
-  header("content-type: application/json");
+if ($_SERVER['REQUEST_METHOD'] === 'GET' && isset($_GET['userId'])) {
   $users = new UsersController();
   echo json_encode($users->getOneUser($_GET["userId"]));
-});
+}
 
-Utils::buildRoute("/users/update/:id", "POST", function () {
+if ($_SERVER['REQUEST_METHOD'] === 'PUT') {
   $users = new UsersController();
   $users->setUserName($_POST["userName"]);
   $users->setUserEmail($_POST["userEmail"]);
   $users->setUserPassword($_POST["userPassword"]);
   echo $users->updateUser($_GET["userId"]);
-});
+}
 
-Utils::buildRoute("/users/delete/:userId", "GET", function () {
+if ($_SERVER['REQUEST_METHOD'] === 'DELETE') {
   $users = new UsersController();
   echo $users->deleteUser($_GET["userId"]);
-});
+}
 ?>
